requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.60.1
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.2.6
packaging==25.0
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0

# Added for FastAPI server
fastapi
uvicorn
python-dotenv
